requirements.txt
diffusers>=0.32.1
accelerate>=1.1.1
transformers>=4.46.2
numpy==1.26.0
torchvision>=0.20.0
sentencepiece>=0.2.0
imageio>=2.35.1
imageio-ffmpeg>=0.5.1
moviepy>=2.0.0
scikit-video>=1.1.11
pydantic>=2.10.3
peft
